requirements.txt
asgiref==3.8.1
Django==5.2.1
django-extensions==4.1
django-taggit==6.1.0
djangorestframework==3.16.0
drf-redesign==0.5.1
sqlparse==0.5.3
